requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anthropic==0.21.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
cchardet==2.1.7
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cssselect==1.2.0
distro==1.9.0
Django==4.2.4
django-apscheduler==0.6.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-comment-migrate==0.1.7
django-cors-headers==4.2.0
django-db==0.0.7
django-debug-toolbar==4.3.0
django-filter==23.2
django-ranged-response==0.2.0
django-rest-swagger==2.2.0
django-restql==0.15.3
django-simple-captcha==0.5.18
django-simpleui==2023.12.12
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
docopt==0.6.2
drf-yasg==1.21.7
drf-yasg2==1.19.4
eventlet==0.33.3
exceptiongroup==1.2.0
fake-useragent==1.2.1
Faker==19.3.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2024.3.1
gevent==23.9.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpath==0.82.2
kombu==5.3.2
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
openai==1.16.2
openapi-codec==1.3.2
packaging==23.1
parse==1.19.1
Pillow==10.0.0
pipreqs==0.4.13
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pyee==8.2.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyPEG2==2.15.2
pyppeteer==1.0.2
pyquery==2.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-html==0.10.0
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
sentry-sdk==1.29.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tokenizers==0.15.2
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.6
websockets==10.4
win32-setctime==1.1.0
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
